Library

Public API

Atmosphere.atmospherefit — Method.

atmospherefit(altitude::Real)

Return density (kg/m^3), viscosity (Pa-s), and speed of sound (m/s) functions.

Input altitude must be in meters.

source

Atmosphere.atmospheretable — Method.

atmospheretable(altitude::Real)

Return density (kg/m^3), viscosity (Pa-s), and speed of sound (m/s) functions.

Input altitude must be in meters.

source

Atmosphere.gravity — Function.

gravity(altitude::Real)

Calculate adjusted gravitational acceleration at altitude (meters) and latitude (degrees).

1984 Ellipsoidal Gravity Formula.

source

Internal API

Atmosphere.density — Method.

density(Temperature::Real, Pressure::Real)

Return Air Density (kg/m^3) from Ideal Gas Law.

Input Temperature must be in Kelvin, and Input Pressure must be in Pascals.

source

Atmosphere.speedofsound — Method.

speedofsound(Temperature::Real)

Return Speed of Sound (m/s) from Ideal Gas Law.

Input Temperature must be in Kelvin.

source

Atmosphere.temp_presfit — Method.

temp_presfit(altitude::Real)

Calculate atmospheric temperature and pressure for input altitude (meters) using fits of the standard atmosphere model, slightly modified from those found in Flight Vehicle Aerodynamics by Mark Drela.

source

Atmosphere.temp_prestable — Method.

temp_prestable(altitude::Real)

Calculate atmospheric temperature and pressure for input altitude (meters) from 1976 Standard Atmosphere Model.

source

Atmosphere.viscosity — Method.

viscosity(Temperature::Real)

Return Dynamic Viscosity (Pa-s) from Sutherlands Equation.

Input Temperature must be in Kelvin.

source